Extract echoing output into an EchoWriter
